refactor(match): 优化比赛创建逻辑并改进数据访问层返回值处理
- 修改 MatchGameRepository.create 方法返回类型从 int 到 Long
- 移除领域服务中多余的 findById 查询操作
- 改进插入失败时的错误处理机制
- 简化比赛ID获取流程提高代码可读性
- 添加 LocalDateTime 导入优化依赖管理

+++ src/main/java/com/csxuhuan/gelatoni/infrastructure/repository/MatchGameRepository.java
@@ -1,8 +1,6 @@
 package com.csxuhuan.gelatoni.infrastructure.repository;
 
 import com.csxuhuan.gelatoni.domain.model.entity.MatchGame;
-import com.csxuhuan.gelatoni.domain.model.entity.MatchTeamStats;
-import com.csxuhuan.gelatoni.domain.model.entity.MatchPlayerStats;
 
 import java.util.List;
 
@@ -47,7 +45,7 @@ public interface MatchGameRepository {
      * @param creator 创建人ID
      * @return 新增的ID
      */
-    int create(MatchGame game, Long creator);
+    Long create(MatchGame game, Long creator);
 
     /**
      * 更新比赛
